// frontend/generated/models/SetDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 8lind8box Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SlotDto } from './SlotDto';
import {
    SlotDtoFromJSON,
    SlotDtoFromJSONTyped,
    SlotDtoToJSON,
    SlotDtoToJSONTyped,
} from './SlotDto';
import type { StockKeepingUnitDto } from './StockKeepingUnitDto';
import {
    StockKeepingUnitDtoFromJSON,
    StockKeepingUnitDtoFromJSONTyped,
    StockKeepingUnitDtoToJSON,
    StockKeepingUnitDtoToJSONTyped,
} from './StockKeepingUnitDto';
import type { BlindBoxDto } from './BlindBoxDto';
import {
    BlindBoxDtoFromJSON,
    BlindBoxDtoFromJSONTyped,
    BlindBoxDtoToJSON,
    BlindBoxDtoToJSONTyped,
} from './BlindBoxDto';

/**
 * 
 * @export
 * @interface SetDto
 */
export interface SetDto {
    /**
     * 
     * @type {number}
     * @memberof SetDto
     */
    setId?: number;
    /**
     * 
     * @type {StockKeepingUnitDto}
     * @memberof SetDto
     */
    sku?: StockKeepingUnitDto;
    /**
     * 
     * @type {boolean}
     * @memberof SetDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Array<SlotDto>}
     * @memberof SetDto
     */
    slots?: Array<SlotDto>;
    /**
     * 
     * @type {BlindBoxDto}
     * @memberof SetDto
     */
    blindBox?: BlindBoxDto;
    /**
     * 
     * @type {Date}
     * @memberof SetDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof SetDto
     */
    updatedAt?: Date;
}

/**
 * Check if a given object implements the SetDto interface.
 */
export function instanceOfSetDto(value: object): value is SetDto {
    return true;
}

export function SetDtoFromJSON(json: any): SetDto {
    return SetDtoFromJSONTyped(json, false);
}

export function SetDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): SetDto {
    if (json == null) {
        return json;
    }
    return {
        
        'setId': json['setId'] == null ? undefined : json['setId'],
        'sku': json['sku'] == null ? undefined : StockKeepingUnitDtoFromJSON(json['sku']),
        'isVisible': json['isVisible'] == null ? undefined : json['isVisible'],
        'slots': json['slots'] == null ? undefined : ((json['slots'] as Array<any>).map(SlotDtoFromJSON)),
        'blindBox': json['blindBox'] == null ? undefined : BlindBoxDtoFromJSON(json['blindBox']),
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
    };
}

export function SetDtoToJSON(json: any): SetDto {
    return SetDtoToJSONTyped(json, false);
}

export function SetDtoToJSONTyped(value?: SetDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'setId': value['setId'],
        'sku': StockKeepingUnitDtoToJSON(value['sku']),
        'isVisible': value['isVisible'],
        'slots': value['slots'] == null ? undefined : ((value['slots'] as Array<any>).map(SlotDtoToJSON)),
        'blindBox': BlindBoxDtoToJSON(value['blindBox']),
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'updatedAt': value['updatedAt'] == null ? undefined : ((value['updatedAt']).toISOString()),
    };
}